Fixed the game for Internet Explorer!
The crash was being caused by IE caching GET requests. It turns out that - for once - IE was (probably) in the right here - it's perfectly acceptable to cache GET requests. Unfortunately this caused an infinite loop issue.
The name problem was fixed with a simple ajax call from the index to set the name (IE doesn't like you modifying hidden form values).